Fixing depreciated onejar-maven-plugin
This removes the onejar-maven-plugin and replaces it with the
maven-assembly-plugin jar-with-dependencies

I used maven 3.5 and java SE 8 8u131 to do my local
build. This change to the pom.xml file fixes the build issue that
occurred for me.

I am not sure if other people had the same issue with their build
hopefully this helps others.